Approved by Ubuntu Phone Apps Jenkins Bot, Michael Spencer.

+++ tests/autopilot/ubuntu_filemanager_app/tests/test_filemanager.py
@@ -42,10 +42,15 @@ class TestFolderListPage(FileManagerTestCase):
             self.main_view.visible, Eventually(Equals(True)))
 
     def _patch_home(self):
+        #create a temporary home for testing purposes
         temp_dir = tempfile.mkdtemp()
-        shutil.copyfile(
-            os.path.expanduser(os.path.join('~', '.Xauthority')),
-            os.path.join(temp_dir, '.Xauthority'))
+        #if the Xauthority file is in home directory
+        #make sure we copy it to temp home, otherwise do nothing
+        xauth = os.path.expanduser(os.path.join('~', '.Xauthority'))
+        if os.path.isfile(xauth):
+            shutil.copyfile(
+                os.path.expanduser(os.path.join('~', '.Xauthority')),
+                os.path.join(temp_dir, '.Xauthority'))
         self.addCleanup(shutil.rmtree, temp_dir)
         patcher = mock.patch.dict('os.environ', {'HOME': temp_dir})
         patcher.start()
